refactor : notion 관련 함수 server 폴더로 모음
